fix (s21-libft) size_t

+++ school21/42cursus/libft/ft_bzero.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-void	ft_bzero(void *s, int len)
+void	ft_bzero(void *s, size_t len)
 {
 	unsigned char	*ptr;
