+++ school21/42cursus/libft/libft.h
@@ -21,19 +21,19 @@ typedef struct	s_list {
 	struct s_list	*next;
 }				t_list;
 
-void			*ft_memset(void *b, int c, int len);
+void			*ft_memset(void *b, int c, size_t len);
 
-void			ft_bzero(void *s, int n);
+void			ft_bzero(void *s, size_t n);
 
-void			*ft_memcpy(void *dst, const void *src, int len);
+void			*ft_memcpy(void *dst, const void *src, size_t len);
 
-void			*ft_memccpy(void *dst, const void *src, int c, int n);
+void			*ft_memccpy(void *dst, const void *src, int c, size_t n);
 
-void			*ft_memmove(void *dst, const void *src, int len);
+void			*ft_memmove(void *dst, const void *src, size_t len);
 
-void			*ft_memchr(const void *s, int c, int n);
+void			*ft_memchr(const void *s, int c, size_t n);
 
-int				ft_memcmp(const void *s1, const void *s2, int n);
+int				ft_memcmp(const void *s1, const void *s2, size_t n);
 
 size_t			ft_strlen(const char *s);
